+++ packages/react/Menu/src/components/MenuFooter/index.ts
@@ -1,0 +1,4 @@
+import { MenuFooter } from "./MenuFooter";
+
+export { MenuFooter } from "./MenuFooter";
+export default MenuFooter;