+++ packages/react/Menu/src/components/MenuFooter/menuFooter.types.ts
@@ -1,0 +1,3 @@
+import { MenuButtonProps } from "@nimbus-ds/menubutton";
+
+export type MenuFooterProps = Omit<MenuButtonProps, "children">;