+++ packages/react/Menu/src/components/MenuSection/menuSection.types.ts
@@ -1,0 +1,6 @@
+import { HTMLAttributes, ReactNode } from "react";
+
+export interface MenuSectionProps extends HTMLAttributes<HTMLElement> {
+  title?: string;
+  children: ReactNode;
+}