+++ packages/react/Menu/src/components/index.ts
@@ -1,0 +1,4 @@
+export * from "./MenuHeader";
+export * from "./MenuBody";
+export * from "./MenuFooter";
+export * from "./MenuSection";