+++ packages/react/MenuButton/src/index.ts
@@ -1,0 +1,5 @@
+import { MenuButton } from "./MenuButton";
+
+export { MenuButton } from "./MenuButton";
+export type { MenuButtonProps } from "./menuButton.types";
+export default MenuButton;